🔨 [TEA-50]: Refactor, changed file structure
